Wait for service annotations to get updated
Step PatchLoadBalancerServicesIPPoolStep updates the
annotation metallb.universe.tf/address-pool for the
service object and expects kubernetes to add anotation
metallb.universe.tf/ip-allocated-from-pool. In some cases,
the value for annotation updated by kubernetes is not the
correct one and it gets refreshed after some time.

Add logic to wait for annotations to get updated to the
proper pool name.

Fixes: https://bugs.launchpad.net/snap-openstack/+bug/2105566