Allow passing initial model configuration
For the model created by Sunbeam, allow passing initial model
configuration. The only guarantee made is that, at creation time, the
parameter will be given to juju, and only if Sunbeam does not override
them. Sunbeam will not perform any checks to see wheter or not these
config options are valid parameter.

Closes-Bug: #2065490
Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>