Determine Coredns replica based on number of control nodes
Currently the HPA applied on coredns has minReplica as 1.
This means in case of multi node deployment, the number
of codedns pods is 1 if there is no load on coredns.

The ideal situation is to have 3 coredns in case of multi
node deployment irrespective of the load. Determine the
number of replicas for coredns based on number of control
nodes.
Apply the determined replica during deploy command in
mass mode and resize command in manual mode.