local: allow specifying "no-physnet"
Right now, the prompt is telling the user to pass an empty
string if no physnet is used. However, the default value
will be picked in this case (e.g. a previously configured
physnet).

The spec says that we should use "no-physnet" in this case,
so we'll update the prompt accordingly. Note that it also
accepts the "null" or "none" keywords.