Adds shared-filesystems feature
OpenStack Manila provides Shared Filesystems as a service.

The manila-k8s charm contains the control plane services, while
the manila-cephfs-k8s charm can integrate with MicroCeph to provide
Manila with the NFS storage backend.

The shared-filesystems (manila) feature requires the microceph's
ceph-nfs relation. That needs to exist before applying the main plan.

This can be resolved by adding a step to the shared-filesystems feature,
in which we create the offer if needed.

Signed-off-by: Claudiu Belu <cbelu@cloudbasesolutions.com>