Keep the previous values in non-interactive mode without a manifest
We'll make a small change to the non-interactive mode. If a manifest
was provided, we'll use that as the main source of truth and drop
the previous answers. If no manifest was provided, we'll keep
the previous settings.

This allows the user to remove one (or all) pci specs in
non-interactive mode, while preserving previous configuration if no
manifest was provided.