Add SAML2 support
This change adds SAML2 support to the identity feature. The SAML2 support
leverages the new keystone-saml-k8s charm to send the required information
to the keystone charm in order for keystone to properly configure apache2.

The SAML2 enablement also requires that the user set a proper x509 certificate
and key in keystone, which will serve as the SP certificate for SAML2.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>