Add DPDK configuration
This PR will allow enabling and configuring OVS DPDK, specifically:

* the number of cores used by the OVS DPDK control plane and data path
* the amount of memory used by OVS DPDK from huge pages
* physical DPDK ports
* DPDK-compatible driver (defaults to vfio-pci)

The specified amount of cores will be reserved in the EPA service,
make sure to configure isolated CPUs using kernel arguments.
At the same time, 1GB huge pages must be preallocated.

The pinned CPUs and memory will be distributed across NUMA nodes based
on the NUMA location of the physical DPDK interfaces.

Note that the specified DPDK interfaces will be persistently bound
to the configured DPDK-compatible driver and will no longer be visible
on the host.

Any physical interfaces that are going to be used with DPDK are
expected to be connected to OVS bridges, either directly or
through a bond. The bridge and bond definitions should be predefined
through Maas/Netplan.

The openstack-hypervisor charm will take care of disabling the system
ovs installation, move bridge definitions to the snap installation
(through netplan) and then reconfigure them to use the newly
defined DPDK ports.

Note that Netplan is not aware of DPDK ports, as such those
will be removed from the Netplan configuration and defined in OVS.

We expect the host Netplan installation to include the following
change in order to work with snap OVS installations:
https://github.com/canonical/netplan/pull/549

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>