sr-iov: allow dropping previous config
When configuring SR-IOV, the new specs are merged with the previous
ones. The reason is that the pci whitelist is shared across the
compute nodes and we want to preserve the specs from other nodes.

The user may need to actually clear the previous config, as such
we're adding a flag that can be specified to drop these settings
instead of merging them with the new specs.