Extend sr-iov functional test
Just like the DPDK functional test, we'll have the SR-IOV test
create a test instance and attach an SR-IOV VF, validating the
resulting Libvirt domain xml.

If the specified interface supports hardware offloading, we'll
enable switchdev mode and use the demo network, which is expected
to be an overlay network.

If hardware offloading is not available, we'll use the public
network instead since in this case overlay networks are unsupported.