Fix EnsureK8SUnitsTaggedStep
Currently the nodes to update take into account
if the label is hostname returned by clusterdb
nodes and hostname in k8s metadata name does not
have domain and so filtering of k8s node fails.

Use both fqdn and hostname without domain in the
checks to filter nodes. However use hostname for
the check with label since the label should always
have hostname same as in clusterdb.

Fixes: https://bugs.launchpad.net/snap-openstack/+bug/2121840